+++ examples/bootstrap.rs
@@ -10,9 +10,9 @@ fn main() {
 
     let client = Dht::client().unwrap();
 
-    client.bootstrapped().unwrap();
+    client.bootstrapped();
 
-    let info = client.info().unwrap();
+    let info = client.info();
 
     println!("{:?}", info);
 }