+++ src/rpc/iterative_query.rs
@@ -179,7 +179,7 @@ impl IterativeQuery {
             .any(|&tid| socket.inflight(&tid));
 
         if done {
-            debug!(id=?self.target(), candidates = ?self.closest.len(), visited = ?self.visited.len(), responders = ?self.responders.len(), "Done query");
+            debug!(id=?self.target(), closest = ?self.closest.len(), visited = ?self.visited.len(), responders = ?self.responders.len(), "Done query");
         };
 
         done